refactor(TypeScript): separat file for Collapse class
* put Collapse in a separat file
* rename tobago-collapsible-operation.ts to match class name
