+++ tobago-theme/tobago-theme-standard/src/main/ts/tobago-collapse-operation.ts
